bot/requirements.txt
anyio==4.9.0
asyncpg==0.30.0
certifi==2025.4.26
charset-normalizer==3.4.2
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.26.0
idna==3.10
python-dotenv==1.1.0
python-telegram-bot==20.8
requests==2.32.3
setuptools==80.7.1
sniffio==1.3.1
SQLAlchemy==2.0.41
typing_extensions==4.13.2
urllib3==2.4.0
wheel==0.45.1
